pyproject.toml
[project]
name = "web-crawler-api"
version = "0.1.0"
description = "Web crawler API with Playwright and Markdown conversion"
authors = [{ name = "Developer", email = "developer@example.com" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "playwright>=1.48.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0.2",
    "markdownify>=0.12.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.0"
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
